fix: move function to separate file and add unit test
